Redirects for docs and community
Redirect /community to the homepage, as it was in the old
[snapcraft.io-static-pages repo](https://github.com/canonical-websites/snapcraft.io-static-pages/).

Redirect /docs to docs.snapcraft.io. This is in preparation for when
docs.snapcraft.io becomes the home of snapcraft docs. It won't effect
anything right now, because [kubernetes takes care of](https://github.com/canonical-webteam/k8s-deployments/blob/master/staging.snapcraft.io/snapcraft.io-ingress.yaml#L27) routing /docs to
the docs application, before it hits this application.