Add X-Commit-ID & X-Hostname headers
These should pick up on the COMMIT_ID environment variable and the hostname of the running context respectively.

The COMMIT_ID variable should exist in the built docker image. The `hostname` should be the name of the pod when the image is running in Kubernetes.

If the COMMIT_ID environment variable isn't found, the HTTP header will still be there, but will be set to None.