Scoping public API calls to specific fields.
Requesting explicit fields will make it easier to migrate to new APIs (similarly to what is already done in snapd).

Transitioning snap-details to `media` (instead of `icon_url` and `screenshot_urls`) which makes it ready for "banner" and "banner_icon" support. Search/promoted/featured-snaps remain using `icon_url` mostly for convenience, but eventually it should also be ported to `media`.