Fix entrypoint logic
We had a problem that when FLASK_DEBUG was unset, `if ${FLASK_DEBUG}` would evaluate to true.

I've fixed this by using the better syntax of `if [ "${FLASK_DEBUG}" = true ]`. For discussion see https://stackoverflow.com/questions/2953646/how-to-declare-and-use-boolean-variables-in-shell-script.

I've also added a "RUN_COMMAND" variable so we can more clearly show the difference between development and production settings.