Update supported openSUSE Leap versions
Fixes #3043

The repository for Leap 42.3 no longer exists.
Leap 42.3 is no longer supported, so this is not unexpected.
The currently available repositories can be found here:
https://download.opensuse.org/repositories/system:/snappy/
Leap 15.1 and 15.2 are now available in addition to 15.0.

This commit removes references to Leap 42.3.
Additionally, it adds references for Leap 15.1 and 15.2.
The example has been updated to use the latest Leap repository, 15.2.