distros: ask debian users to install core snap
This brings the "Enable snapd" instructions in line with what is
documented at https://snapcraft.io/docs/installing-snap-on-debian.

The versions of snapd that ships with Debian stable and oldstable do not
fully support base snaps, which is required to successfully install many
snaps.

This commit addresses that issue by asking the user to upgrade snapd as
part of the installation process.